+++ src/functions/emit/diff_generator.c
@@ -3,15 +3,16 @@
  * @brief Implementation of CST diff generation.
  */
 
-/* clang-format off */
+/* clang-format off */#include "c_cdd/safe_crt_msvc.h"
+
 #include "c_cdd_export.h"
 #include <errno.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
-#include "functions/emit/diff_generator.h"
 #include "c_cdd/log.h"
+#include "functions/emit/diff_generator.h"
 /* clang-format on */
 
 cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
@@ -51,11 +52,11 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
 /* Print header */
 #if defined(_MSC_VER) && !defined(__INTEL_COMPILER) ||                         \
     defined(__STDC_LIB_EXT1__) && __STDC_WANT_LIB_EXT1__
-  diff_len += sprintf_s(diff_buf + diff_len, diff_cap - diff_len,
-                        "--- a/%s\n+++ b/%s\n", filename, filename);
+  diff_len += (size_t)sprintf_s(diff_buf + diff_len, diff_cap - diff_len,
+                                "--- a/%s\n+++ b/%s\n", filename, filename);
 #else
-  diff_len +=
-      sprintf(diff_buf + diff_len, "--- a/%s\n+++ b/%s\n", filename, filename);
+  diff_len += (size_t)sprintf(diff_buf + diff_len, "--- a/%s\n+++ b/%s\n",
+                              filename, filename);
 #endif
   for (i = 0; i < list->size; ++i) {
     const struct Patch *p = &list->patches[i];
@@ -63,10 +64,11 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
     /* This is a simplification; we'll output a chunk */
 #if defined(_MSC_VER) && !defined(__INTEL_COMPILER) ||                         \
     defined(__STDC_LIB_EXT1__) && __STDC_WANT_LIB_EXT1__
-    diff_len += sprintf_s(diff_buf + diff_len, diff_cap - diff_len,
-                          "@@ -patch %d @@\n", (int)i);
+    diff_len += (size_t)sprintf_s(diff_buf + diff_len, diff_cap - diff_len,
+                                  "@@ -patch %d @@\n", (int)i);
 #else
-    diff_len += sprintf(diff_buf + diff_len, "@@ -patch %d @@\n", (int)i);
+    diff_len +=
+        (size_t)sprintf(diff_buf + diff_len, "@@ -patch %d @@\n", (int)i);
 #endif
 
     /* Emit old tokens (we prefix with '-') */
@@ -131,7 +133,7 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
     diff_len +=
         sprintf_s(diff_buf + diff_len, diff_cap - diff_len, "+%s\n", p->text);
 #else
-    diff_len += sprintf(diff_buf + diff_len, "+%s\n", p->text);
+    diff_len += (size_t)sprintf(diff_buf + diff_len, "+%s\n", p->text);
 #endif
   }
